+++ netket/__init__.py
@@ -48,7 +48,6 @@ from . import (
     _exact_dynamics,
     _vmc,
     _steadystate,
-    _dynamics,
 )
 
 # Main applications
